Extract createConverter
Extract createConverter into file to help with future enhancements.
