+++ src/basic-types.ts
@@ -1,4 +1,5 @@
-import { Converter, createConverter } from './core';
+import { Converter } from './core';
+import { createConverter } from './create-converter';
 import { ConverterError } from './errors';
 import { displayValue } from './formatting';
