+++ src/objects.ts
@@ -1,4 +1,5 @@
-import { Converter, ConverterFunction, createConverter, getConverterName } from './core';
+import { Converter, ConverterFunction, getConverterName } from './core';
+import { createConverter } from './create-converter'; 
 import { optional } from './decorators';
 import { ConverterError } from './errors';
