+++ src/paths.ts
@@ -1,5 +1,6 @@
 import { unknown as UnknownConverter } from './basic-types';
-import { Converter, ConverterFunction, createConverter, getConverterName } from './core';
+import { Converter, ConverterFunction, getConverterName } from './core';
+import { createConverter } from './create-converter';
 import { ConverterError } from './errors';
 import { formatPath } from './formatting';
